[Fix #116] fix padding_values not initialized when using FixedLenSequenceFeature
When I use `tfr.data.read_batched_sequence_example_dataset` with example_feature_spec containing `tf.FixedLenSequenceFeature`, it raises a KeyError.

In `_SequenceExampleParser.parse`,  it convert `FixedLenFeature` to `FixedLenSequenceFeature` and setting dict `padding_values` at the same time.

But when I pass FixedLenSequenceFeature directly, `padding_values` is not set, which remains empty. And in the following code calling `pad_fn` it raises KeyError.

Changing `FixedLenSequenceFeature ` to `FixedLenFeature ` is OK, but I think it's counter-intuitive. It should be compatible with `FixedLenSequenceFeature ` .